+++ src/assets/Core.php
@@ -35,24 +35,24 @@ JS
 
     public $css = [];
 
-    /*public $js = [
+    public $js = [
         'js/Skeeks.js',
         'js/Classes.js',
         'js/Entity.js',
         'js/EventManager.js',
         'js/Config.js',
-        'js/Cookie.js',
         'js/Component.js',
+        'js/Cookie.js',
         'js/Ajax.js',
-    ];*/
+    ];
 
     /**
      * @see http://closure-compiler.appspot.com/home
      * @var array
      */
-    public $js = [
+    /*public $js = [
         'distr/skeeks-core.min.js',
-    ];
+    ];*/
 
     public $depends = [
         'yii\web\YiiAsset',